combined all electrode builders into one file
